+++ letsencrypt-win-simple/Certificates/CertificateService.cs
@@ -1,0 +1,141 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Security.Cryptography;
+using System.Security.Cryptography.X509Certificates;
+using LetsEncrypt.ACME.Simple.Configuration;
+using Serilog;
+
+namespace LetsEncrypt.ACME.Simple.Certificates
+{
+    public class CertificateService
+    {
+        public void InstallCertificate(Target binding, string pfxFilename, out X509Store store,
+            out X509Certificate2 certificate)
+        {
+            try
+            {
+                store = new X509Store(App.Options.CertificateStore, StoreLocation.LocalMachine);
+                store.Open(OpenFlags.OpenExistingOnly | OpenFlags.ReadWrite);
+            }
+            catch (CryptographicException)
+            {
+                store = new X509Store(StoreName.My, StoreLocation.LocalMachine);
+                store.Open(OpenFlags.OpenExistingOnly | OpenFlags.ReadWrite);
+            }
+            catch (Exception ex)
+            {
+                Log.Error("Error encountered while opening certificate store. Error: {@ex}", ex);
+                throw new Exception(ex.Message);
+            }
+
+            Log.Information("Opened Certificate Store {Name}", store.Name);
+            certificate = null;
+            try
+            {
+                X509KeyStorageFlags flags = X509KeyStorageFlags.MachineKeySet | X509KeyStorageFlags.PersistKeySet;
+                if (Properties.Settings.Default.PrivateKeyExportable)
+                {
+                    Console.WriteLine($" Set private key exportable");
+                    Log.Information("Set private key exportable");
+                    flags |= X509KeyStorageFlags.Exportable;
+                }
+
+                // See http://paulstovell.com/blog/x509certificate2
+                certificate = new X509Certificate2(pfxFilename, Properties.Settings.Default.PFXPassword,
+                    flags);
+
+                certificate.FriendlyName =
+                    $"{binding.Host} {DateTime.Now.ToString(Properties.Settings.Default.FileDateFormat)}";
+                Log.Debug("{FriendlyName}", certificate.FriendlyName);
+
+                Log.Information("Adding Certificate to Store");
+                store.Add(certificate);
+
+                Log.Information("Closing Certificate Store");
+            }
+            catch (Exception ex)
+            {
+                Log.Error("Error saving certificate {@ex}", ex);
+            }
+            store.Close();
+        }
+
+        public void UninstallCertificate(string host, out X509Store store, X509Certificate2 certificate)
+        {
+            try
+            {
+                store = new X509Store(App.Options.CertificateStore, StoreLocation.LocalMachine);
+                store.Open(OpenFlags.OpenExistingOnly | OpenFlags.ReadWrite);
+            }
+            catch (CryptographicException)
+            {
+                store = new X509Store(StoreName.My, StoreLocation.LocalMachine);
+                store.Open(OpenFlags.OpenExistingOnly | OpenFlags.ReadWrite);
+            }
+            catch (Exception ex)
+            {
+                Log.Error("Error encountered while opening certificate store. Error: {@ex}", ex);
+                throw new Exception(ex.Message);
+            }
+
+            Log.Information("Opened Certificate Store {Name}", store.Name);
+            try
+            {
+                X509Certificate2Collection col = store.Certificates.Find(X509FindType.FindBySubjectName, host, false);
+
+                foreach (var cert in col)
+                {
+                    var subjectName = cert.Subject.Split(',');
+
+                    if (cert.FriendlyName != certificate.FriendlyName && subjectName[0] == "CN=" + host)
+                    {
+                        Log.Information("Removing Certificate from Store {@cert}", cert);
+                        store.Remove(cert);
+                    }
+                }
+
+                Log.Information("Closing Certificate Store");
+            }
+            catch (Exception ex)
+            {
+                Log.Error("Error removing certificate {@ex}", ex);
+            }
+            store.Close();
+        }
+        
+        public void GetCertificateForTargetId(List<Target> targets, int targetId)
+        {
+            if (!App.Options.San)
+            {
+                var targetIndex = targetId - 1;
+                if (targetIndex >= 0 && targetIndex < targets.Count)
+                {
+                    Target binding = GetBindingByIndex(targets, targetIndex);
+                    binding.Plugin.Auto(binding);
+                }
+            }
+            else
+            {
+                Target binding = GetBindingBySiteId(targets, targetId);
+                binding.Plugin.Auto(binding);
+            }
+        }
+
+        private static Target GetBindingByIndex(List<Target> targets, int targetIndex)
+        {
+            return targets[targetIndex];
+        }
+
+        private static Target GetBindingBySiteId(List<Target> targets, int targetId)
+        {
+            return targets.First(t => t.SiteId == targetId);
+        }
+        
+        public void GetCertificatesForAllHosts(List<Target> targets)
+        {
+            foreach (var target in targets)
+                target.Plugin.Auto(target);
+        }
+    }
+}
